backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.4.0
redis==5.0.8
django-celery-beat==2.6.0
africastalking==1.2.5
phonenumbers==8.13.39
Pillow==10.4.0
gunicorn==22.0.0
